add integration_test to pubspec and turn `add_ev_test` into an integration test (the test actually works now)
