+++ ev_charge/pubspec.yaml
@@ -44,6 +44,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  integration_test:
+    sdk: flutter
   drift_dev: ^2.26.0
   build_runner: ^2.4.15
